refactor: rename API to rpc, and classes from `{X}API` to `{X}RPCAPI` to clarify naming.
